Convert Queue & Queue test to TS
